Reorganize existing examples.
Give existing examples a more regular format.
