+++ mlops_manager/templates/app/examples/fashion-mnist/__APP_NAME__/__APP_NAME__/pipelines/fashion_mnist/pipeline.py
@@ -1,5 +1,5 @@
 # TODO(axelmagn): delete in favor of standalone pipeline template
-from ..config import get_config
+from ...config import get_config
 from kfp import dsl
 from kfp.v2.dsl import (Dataset, Input, Metrics, Model, Output, component)
