+++ mlops_manager/templates/app/examples/fashion-mnist/__APP_NAME__/__APP_NAME__/trainers/fashion_mnist/task.py
