+++ mlops_manager/templates/app/examples/fashion-mnist/__APP_NAME__/config/fashion_mnist.yaml
@@ -1,0 +1,4 @@
+pipelines:
+  fashion_mnist:
+    build:
+      python_path: {{app_name}}.pipelines.fashion_mnist.pipeline.pipeline
