+++ mlops_manager/templates/pipeline/kfp/__APP_NAME__/__APP_NAME__/pipelines/__PIPELINE_NAME__/pipeline.py
